move mk* functions from resolver to UnkindedType
