+++ main/src/ca/uwaterloo/flix/language/phase/Resolver.scala
@@ -17,7 +17,8 @@
 package ca.uwaterloo.flix.language.phase
 
 import ca.uwaterloo.flix.api.Flix
-import ca.uwaterloo.flix.language.ast.Ast.{BoundBy, Denotation}
+import ca.uwaterloo.flix.language.ast.Ast.BoundBy
+import ca.uwaterloo.flix.language.ast.UnkindedType.{mkAnd, mkComplement, mkEffect, mkEnum, mkIntersection, mkNot, mkOr, mkPredicate, mkUnappliedTypeAlias, mkUncurriedArrowWithEffect, mkUnion}
 import ca.uwaterloo.flix.language.ast.{Symbol, _}
 import ca.uwaterloo.flix.language.errors.ResolutionError
 import ca.uwaterloo.flix.util.Validation._
@@ -1779,7 +1780,7 @@ object Resolver {
       }
 
     case NamedAst.Type.Enum(sym, loc) =>
-      mkUnkindedEnum(sym, loc).toSuccess
+      mkEnum(sym, loc).toSuccess
 
     case NamedAst.Type.Tuple(elms0, loc) =>
       val elmsVal = traverse(elms0)(tpe => semiResolveType(tpe, ns0, root))
@@ -2387,7 +2388,7 @@ object Resolver {
     */
   private def getEnumTypeIfAccessible(enum0: NamedAst.Enum, ns0: Name.NName, loc: SourceLocation): Validation[UnkindedType, ResolutionError] =
     getEnumIfAccessible(enum0, ns0, loc) map {
-      case enum => mkUnkindedEnum(enum.sym, loc)
+      case enum => mkEnum(enum.sym, loc)
     }
 
   /**
@@ -2689,95 +2690,9 @@ object Resolver {
   }
 
   /**
-    * Returns a synthetic namespace obtained from the given sequence of namespace `parts`.
-    */
-  private def getNS(parts: List[String]): Name.NName = {
-    val sp1 = SourcePosition.Unknown
-    val sp2 = SourcePosition.Unknown
-    val idents = parts.map(s => Name.Ident(sp1, s, sp2))
-    Name.NName(sp1, idents, sp2)
-  }
-
-  /**
-    * Construct the enum type constructor for the given symbol `sym` with the given kind `k`.
-    */
-  def mkUnkindedEnum(sym: Symbol.EnumSym, loc: SourceLocation): UnkindedType = UnkindedType.Enum(sym, loc)
-
-  /**
-    * Construct the enum type `Sym[ts]`.
-    */
-  def mkUnkindedEnum(sym: Symbol.EnumSym, ts: List[Symbol.UnkindedTypeVarSym], loc: SourceLocation): UnkindedType = {
-    val args = ts.map(sym => UnkindedType.Var(sym, sym.loc))
-    UnkindedType.mkApply(UnkindedType.Enum(sym, loc), args, loc)
-  }
-
-  /**
     * Construct the type alias type constructor for the given symbol `sym` with the given kind `k`.
     */
   def mkUnappliedTypeAlias(sym: Symbol.TypeAliasSym, loc: SourceLocation): UnkindedType = UnkindedType.UnappliedAlias(sym, loc)
-
-  /**
-    * Construct the effect type for the given symbol.
-    */
-  def mkEffect(sym: Symbol.EffectSym, loc: SourceLocation): UnkindedType = UnkindedType.Cst(TypeConstructor.Effect(sym), loc)
-
-  /**
-    * Constructs a predicate type.
-    */
-  private def mkPredicate(den: Ast.Denotation, ts0: List[UnkindedType], loc: SourceLocation): UnkindedType = {
-    val tycon = den match {
-      case Denotation.Relational => UnkindedType.Cst(TypeConstructor.Relation, loc)
-      case Denotation.Latticenal => UnkindedType.Cst(TypeConstructor.Lattice, loc)
-    }
-    val ts = ts0 match {
-      case Nil => UnkindedType.Cst(TypeConstructor.Unit, loc)
-      case x :: Nil => x
-      case xs => UnkindedType.mkTuple(xs, loc)
-    }
-
-    UnkindedType.Apply(tycon, ts, loc)
-  }
-
-  /**
-    * Returns the type `Not(tpe1)`.
-    */
-  private def mkNot(tpe1: UnkindedType, loc: SourceLocation): UnkindedType = UnkindedType.mkApply(UnkindedType.Cst(TypeConstructor.Not, loc), List(tpe1), loc)
-
-  /**
-    * Returns the type `And(tpe1, tpe2)`.
-    */
-  private def mkAnd(tpe1: UnkindedType, tpe2: UnkindedType, loc: SourceLocation): UnkindedType = UnkindedType.mkApply(UnkindedType.Cst(TypeConstructor.And, loc), List(tpe1, tpe2), loc)
-
-  /**
-    * Returns the type `Or(tpe1, tpe2)`.
-    */
-  private def mkOr(tpe1: UnkindedType, tpe2: UnkindedType, loc: SourceLocation): UnkindedType = UnkindedType.mkApply(UnkindedType.Cst(TypeConstructor.Or, loc), List(tpe1, tpe2), loc)
-
-  /**
-    * Returns the type `Complement(tpe1)`.
-    */
-  private def mkComplement(tpe1: UnkindedType, loc: SourceLocation): UnkindedType = UnkindedType.mkApply(UnkindedType.Cst(TypeConstructor.Complement, loc), List(tpe1), loc)
-
-  /**
-    * Returns the type `Union(tpe1, tpe2)`.
-    */
-  private def mkUnion(tpe1: UnkindedType, tpe2: UnkindedType, loc: SourceLocation): UnkindedType = UnkindedType.mkApply(UnkindedType.Cst(TypeConstructor.Union, loc), List(tpe1, tpe2), loc)
-
-  /**
-    * Returns the type `Intersection(tpe1, tpe2)`.
-    */
-  private def mkIntersection(tpe1: UnkindedType, tpe2: UnkindedType, loc: SourceLocation): UnkindedType = UnkindedType.mkApply(UnkindedType.Cst(TypeConstructor.Intersection, loc), List(tpe1, tpe2), loc)
-
-  /**
-    * Constructs the uncurried arrow type (A_1, ..., A_n) -> B & e.
-    */
-  def mkUncurriedArrowWithEffect(as: List[UnkindedType], e: UnkindedType.PurityAndEffect, b: UnkindedType, loc: SourceLocation): UnkindedType = {
-    val arrow = UnkindedType.Arrow(e, as.length + 1, loc)
-    val inner = as.foldLeft(arrow: UnkindedType) {
-      case (acc, x) => UnkindedType.Apply(acc, x, loc)
-    }
-    UnkindedType.Apply(inner, b, loc)
-  }
 
   /**
     * Returns either the explicit region (if present), the current region (if present), or the global region.
